src/requirements.txt
pillow==10.3.0
opencv-python==4.10.0.84
pandas==2.2.2
pydantic-settings==2.2.1
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
matplotlib==3.9.0
scikit-learn==1.5.0
transformers==4.41.2
uvicorn==0.30.1
fastapi==0.111.0
enum==0.4.7
aiofiles==24.1.0
motor==3.5.1
datasets==2.20.0
starlette==0.38.1
